requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
chromadb
google-generativeai
pandas
